fix: update revtype column type from TINYINT to SMALLINT in audit tables and add Flyway plugin configuration

+++ src/main/resources/db/migration/V4__add_envers_audit_tables.sql
@@ -6,7 +6,7 @@ CREATE TABLE revinfo (
 CREATE TABLE customers_aud (
     id BIGINT NOT NULL,
     rev INTEGER NOT NULL,
-    revtype TINYINT,
+    revtype SMALLINT,
     full_name VARCHAR(255),
     email VARCHAR(255),
     phone VARCHAR(50),
@@ -20,7 +20,7 @@ CREATE TABLE customers_aud (
 CREATE TABLE leads_aud (
     id BIGINT NOT NULL,
     rev INTEGER NOT NULL,
-    revtype TINYINT,
+    revtype SMALLINT,
     customer_id BIGINT,
     vehicle_id BIGINT,
     dealership_id BIGINT,
@@ -38,7 +38,7 @@ CREATE TABLE leads_aud (
 CREATE TABLE user_type_aud (
     id VARCHAR(255) NOT NULL,
     rev INTEGER NOT NULL,
-    revtype TINYINT,
+    revtype SMALLINT,
     type VARCHAR(255),
     PRIMARY KEY (id, rev),
     CONSTRAINT fk_user_type_aud_rev FOREIGN KEY (rev) REFERENCES revinfo(rev)
@@ -47,7 +47,7 @@ CREATE TABLE user_type_aud (
 CREATE TABLE users_aud (
     id VARCHAR(255) NOT NULL,
     rev INTEGER NOT NULL,
-    revtype TINYINT,
+    revtype SMALLINT,
     username VARCHAR(255),
     email VARCHAR(255),
     hashed_password VARCHAR(255),
